Day_03/health_wellness_agent/pyproject.toml
[project]
name = "health-wellness-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
